requirements.txt
fastapi
uvicorn
llama-cpp-python
# git+https://github.com/abdeladim-s/pyllamacpp.git